+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=a02946560c08de8e678f",
+    "/js/app.js": "/js/app.js?id=4ec0a1971419de7950db",
     "/css/app.css": "/css/app.css?id=35978cdc6a38f7d64fbb"
 }
